Use lhttpc:header_value instead of proplists
HTTP headers are supposed to be case-insensitive, but this is not
respected if we use proplists:get_value to fetch headers.

If we use the patched version of lhttpc in erlcloud 0.4.1 this doesn't
matter since (apparently?) all headers are converted to lowercase.
However, with other versions of lhttpc, we risk not finding header
values due to case mismatches.